Convert to ocamlrep by reference rather than taking ownership
Summary:
Currently, the OcamlRep trait requires a method `into_ocamlrep`, which takes ownership of `self` and deallocates it by letting it fall out of scope. This allows us to free the memory used by the receiver while we are converting it, somewhat reducing peak memory usage.

However, it prevents us from converting values pointed to by an `Rc`, since we cannot free the value as we are converting it if other references exist. We must also provide a method `to_ocamlrep` which takes `self` by reference.

I am not sure how to provide both `into_ocamlrep` and `to_ocamlrep` without lots of code duplication, so I have chosen here to provide `to_ocamlrep` only. We can design a way to provide `into_ocamlrep` in the future if we find that we have a strong need for it.

Reviewed By: losvald

Differential Revision: D17864956

fbshipit-source-id: 7441f637a8f35b511b82d3e91e14f954d8ba6390