Abstract over backing allocator
Summary: I would like ocamlrep to be able to allocate values directly into the OCaml heap via ocamlpool. To do this, the OcamlRep trait needs to abstract over the allocator implementation, so that we can substitute out Arena for an allocator which uses ocamlpool. This change adds an Allocator trait for this purpose.

Reviewed By: ljw1004

Differential Revision: D17864960

fbshipit-source-id: 6ced64bc34a943f3f5e02d9794f07bbf53e46360