Autocomplete for shapes should detect automatically closed apostrophes
Summary:
As reported in workplace posts, VS Code auto-closes single apostrophes.  This apostrophe can mess with shape-key autocomplete; it leads to situations where closing quotes get doubled up:

```
$x = $myshape['x''];
```

This diff adds a test that ensures that closing quotes after the autocomplete position are detected.  When we detect an apostrophe after the closing position, we avoid sending the closing apostrophe in the autocomplete text.

Reviewed By: ljw1004

Differential Revision: D18016344

fbshipit-source-id: 1caebc668915a21016aae93ccba61abbfc58e0c6