Better fix for altered store address
Summary:
D15593915 was a correct, but pessimistic fix for this issue. First, it
marked the store altered at the point that its dependent-address was
defined. This was pessimistic because often its "defined" by something
like LdPropAddr, when what we care about is where the "base" address
was defined. In addition, we only need to mark stores altered at the
start of such a block (since any back-edge would have to go there, or
earlier).

This fixes it to search for a producesReference instruction, or a
DefLabel, and then marks the store altered at the start of that block.

Reviewed By: ricklavoie

Differential Revision: D18045379

fbshipit-source-id: 885b78870ce4e196704c7d54c63a422b00fa08e7