Fix bug in generic LdIterBase
Summary:
An iterator base is a tagged pointer - either an ArrayData (lowest bit 0) or an ObjectData (lowest bit 1). The type byte isn't stored, so to do a generic LdIterBase, we need to mask off the tag bit and then use the HeaderKind to compute a DataType.

The new code adds quite a few more instructions to LdIterBase. That's okay, for two reasons. First, this generic version is only used in profiling translations. Second, we can trivially change cgProfileDecRef in irlower-refcount to avoid using the type byte given that this temp is known to be ref-counted, causing most of the new code to get DCE-d. The only change to the generated code is the extra and. We'll do the DCE change in the next diff, which actually does the profiling.

Reviewed By: jjgriego

Differential Revision: D18038343

fbshipit-source-id: a3630b68d8469ef967b40cdb276f461b9da6a9f2