Reuse prepareAndCallKnown() unpack handling when inlining
Summary:
Run the argument packing code in prepareAndCallKnown() before inlining, so that
all the duplicate logic can be removed.

Make sure the unpack arg is properly handled for cost calculation purposes.
Before this diff, we always replaced it with an empty `varray[]`, which has
probably caused overinlining, as methods that accept unpack args are likely to
iterate them.

Reviewed By: ricklavoie

Differential Revision: D18078586

fbshipit-source-id: 626122a165ad788bb42b90a4e3fb808c58da100a