Delete static-str tracking optimization
Summary:
The "static string optimization" for MixedArrayKeys is where we implicitly assume that all MixedArrays include a static string key. (A large majority of MixedArrays are used as "structs" with static string keys.) With this optimization, we can avoid touching MixedArrayKeys for static string insertions. However, if we keep this optimization on, we can't use this bitset to do an O(1) check that a MixedArray has all int keys.

The final round of performance tests for array iterators didn't show a significant difference between turning int-key specialization on and leaving it off. Since turning on static string tracking can help us with future optimizations, we opted to turn it on and thus enable int-key specialization.

The static string optimization is controlled by a single flag. Now that it's on, we can run a manual DCE over the code.

Reviewed By: jjgriego

Differential Revision: D18045269

fbshipit-source-id: a9674161112fcf85063e94c7bf774283da269f17