Add test for workspace symbol
Summary:
This diffstack fixes issues discovered with server search or command line search.

Since this was broken and only caught by a human being noticing it, let's add an LSP test to verify that it works in the future.

The specific case we need to exercise is searching for a function within a namespace.

Reviewed By: arxanas

Differential Revision: D18076361

fbshipit-source-id: ed6bfc0bba28c0e0e6df612995aede669ea9a768