Convert bytes rather than string slices
Summary: When originally implementing the conversion logic for strings, `&str` seemed like a natural choice. However, a PathBuf can contain invalid UTF-8. There's no reason we shouldn't be able to convert paths which aren't valid UTF-8--after all, as far as the OCaml runtime is concerned, a string is just a byte sequence.

Reviewed By: jewelpit

Differential Revision: D18046138

fbshipit-source-id: 9099d6c3a526dbbe94fc3188c0e224e607e1cf95