Implement OcamlRep for SourceText
Summary:
I'd like to investigate using `ocamlrep` for parser FFI now that it can allocate into ocamlpool using the allocator in {D18045185}. To do that, it would be convenient to have an implementation of `OcamlRep` for `SourceText`.

Instead of caching the converted OCaml source text using `RcOc`, this implementation just uses the `ocaml_source_text` field. It seems that there is already an expectation that a `SourceText` refers to (and must not outlive) an OCaml source text.

Reviewed By: 2BitSalute

Differential Revision: D18051115

fbshipit-source-id: 246653dc0bb29b80ade3bdff7f4808646d53657a