Provide FFI convenience macro for ocamlrep_ocamlpool
Summary:
The ocaml::caml macro isn't a great fit for our needs. Since we allocate most values using ocamlrep::Arena or ocamlpool, and generally do not call into OCaml while holding on to those values, there is no need to root them for the garbage collector. For this reason, I removed the caml macro in some test bindings in {D17968158}.

Even with the `ocamlrep_ocamlpool::to_ocaml` convenience function, we'll still end up writing a fair amount of boilerplate for each FFI function. We need to convert each parameter into a Rust value, then we need to convert the result into an OCaml value. A macro could get rid of some of this boilerplate.

Reviewed By: jewelpit

Differential Revision: D18045185

fbshipit-source-id: 54f024822fc871dc2b5e82de4965a1eedb427e7b