Back out "Delete static-str tracking optimization"
Summary:
Original commit changeset: a9674161112f

Apparently, there's some way to put a static string into a MixedArray without setting this bit. Since we check the bitset's integrity before releasing a MixedArray, missing a key type triggers dbgo assertion failures. We also really don't want to be in that state, as it can cause us to incorrectly pass certain type checks.

Note that this diff also backs out one line from the iterator specialization diff, the one that actually changes kTrackStaticStrKeys to true.

The rest of the iterator specialization diff is independent of this line; I ran Swayzes both with and without it.

Reviewed By: ottoni

Differential Revision: D18089531

fbshipit-source-id: 1f0410d388b7d3af5d0b341a125435501a7cdabb