Clean up ExitPlaceholder and LICM
Summary:
The LICM pass hasn't been functional for a while; a bunch of assertions that it makes about the instructions at a loop's entry are no longer true. (I tried it out a couple months ago; I was able to get some debug from loop-analysis to print by commenting out all mutating operations on the CFG, but the actual hoisting assert-fails in many cases.)

I repurposed the placeholder idea from LICM for iterator specialization. This use case is better described as a "JmpPlaceholder" rather than an "ExitPlaceholder".

Reviewed By: jjgriego

Differential Revision: D18045063

fbshipit-source-id: a6ba0fe0565db799331c5cbd9932a22dc5a8eff5