Fix spill weight calculation for phis in vasm-graph-color
Summary:
When we enter a block to calculate spill weight in it, we remove any
Vregs which aren't live-in to the block. However, when dealing with
phis, we add the new Vregs at the phijmp. Since Vregs defined by a phi
aren't live-in to the block (they get defined by the phidef), they
always get removed. As a result, we effectively always remove Vregs
defined by a phi immediately and don't calculate any weight for
them. The fix is simple. If the block begins with a phidef, don't
remove any Vregs defined by it.

Reviewed By: ottoni

Differential Revision: D18093174

fbshipit-source-id: 8ca50a3dfd82607a6294e4a6899d893d2d03828c