Elaborate New Classes and Records
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.
The plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

Elaborates all New classes as well as records. This concludes the final uses of make_class_id_without_elaboration as all callsites of the old `make_class_id` no longer require elaboration in naming. Thus we can delete the temporary holdover function.

Reviewed By: kmeht

Differential Revision: D17847132

fbshipit-source-id: bfc6d7088d612af1bc255707e63ba8dd7c9f8ccc