Elaborate user attributes
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.

The plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

This diff elaborates `Id` nodes in user_attributes. It also adds `type_params` into the environment but it is not necessary that they are set yet for user attributes it seems.

Reviewed By: kmeht

Differential Revision: D17846690

fbshipit-source-id: 70c138484be4795a527f8c70cac86649e53b4d6c