Set the correct type parameters in the env
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.
The plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

This diff correctly sets the type_params in the endo environment.

Reviewed By: kmeht

Differential Revision: D17847105

fbshipit-source-id: 5e0decfc0a28b76e502c0c9895d7d567788dc43d