Elaborate class consts
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.
The plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

This diff elaborates Id nodes for class_consts.

I created a temporary function `make_class_id_without_elaboration` that is the same as `make_class_id` only it does not elaborate ids. This is used temporarily to verify behavior is correct for testing and deleted at the end of the stack when all uses of make_class_id no longer elaborate ids.

Reviewed By: kmeht

Differential Revision: D17847116

fbshipit-source-id: 127989dffe0f69bbfa2c8764bf0af6513fe8bed2