Remove unnecessary elaboration for unknown record id
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.
The plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

Deletes a use of elaboration for records that would have unknown

Reviewed By: kmeht

Differential Revision: D17847137

fbshipit-source-id: 830f36599a60baa382b9793573ae8ef49ca3fd7e