Extract namespace elaboration into it's own endo pass
Summary:
We currently perform namespace elaboration all over the place - such as in naming and in codegen. If we are able to extract the namespace elaboration into one pass, we can remove having separate logic in naming and codegen and have it all in one place.

This diff is the first in a series of diffs - the plan is to
1.) Create an endo visitor that performs just namespace elaboration
2.) Move all of the elaboration from naming into it
3.) Add this pass into the codegen pass and see if codegen produces different bytecode

This diff creates the endo visitor and tests removing the fun_id_special callsite of elaborate naming from naming.ml

There are lots of issues with moving checks to nast_check. It turns out that naming does not actually name the body of methods and functions (for a reason I do not know why). As a result, we perform nast_checks on the unnamed bodies on functions and methods (I have no idea why this happens). This means, I can no longer assume that the conversion that happens in naming happens before nast_check. So there's duplicated logic there. Which is silly.

Reviewed By: kmeht, shiqicao

Differential Revision: D17755662

fbshipit-source-id: 71d5a37d2b911fdd7945c2a51d524e294f281df7