Support ignoring requests in tests
Summary: We currently support ignored notifications of certain kinds in our LSP tests. Add the same functionality for requests, so that we can use it for non-deterministic `window/showStatus` request. (Note that not *every* such request is non-deterministic, but there may be intermediate ones while it's processing that we want to ignore.)

Reviewed By: 2BitSalute

Differential Revision: D18068088

fbshipit-source-id: d019553a53a8606d43c76f92be1471f95bc21ed4