Fix assert message when spilling when not allowed
Summary:
In certain contexts the register allocator is not allowed to
spill. We check for this and emit a detailed assertion message when
this happens. However we also check this when we check invariants in
the spiller state, using a normal assertion. In one situation, we
accidently will check invariants before doing the specific "can spill"
test. This means you get the generic assertion failure rather than the
intended one. Fix this by moving the check up before dropping dead
registers.

Reviewed By: mofarrell

Differential Revision: D18110605

fbshipit-source-id: 14c814ae95974c3572cf3d05ee29c31eb06056ea