Initialize not passed generics to empty array instead of uninit for the unwinder
Summary: Earlier I fixed this same problem in the JIT on D17750442. With the JIT unwinder, if the uninit is pushed by the interpreter but return instruction is being executed by the JIT, we try to dec ref the uninit as an array which is wrong and causes segfault. This behavior does not happen on master but we should preemptively fix it anyway.

Reviewed By: alexeyt

Differential Revision: D18074322

fbshipit-source-id: a9a74a732295c6cf2c66ecf36a9b0a432c4d0731