Speed up serverless IDE incremental parsing of files
Summary:
The original process for updating the SymbolIndex was to use the FileInfo.t object that was provided to it by the Search Service Runner.  As a result, most of the process for updating the symbol index was designed to use FileInfo.t objects.

Over time we discovered that the FileInfo.t object didn't contain everything the symbol index needed, so we began to use the fast facts parser to rescan the file and gather the extra data we needed.

In the case of serverless IDE, sIDE was using the fast facts parser to scan each incrementally changed file, but then it converted the data to a FileInfo.t object in order to hand it off to the symbol index's updater, which then re-scanned the file via fast-facts-parser.  Yuck!

Let's short circuit this process and just pass the raw facts data through the pipeline.

Reviewed By: 2BitSalute

Differential Revision: D18119859

fbshipit-source-id: 87174a4e87ab0fc1b9142e81f03374aadab41f6c