Fix width analysis in vasm-jumps
Summary:
When trying to convert a diamond into a cmov, we do a width analysis
of the involved Vregs to determine the correctly sized cmov to
use. However, the analysis did not treat physical register and
constant Vregs any differently, which meant it would refine the width
for those Vregs. So, a use of a constant elsewhere in the unit would
affect the width of the cmov. This, combined with the fact that we
didn't consider the phidef def, meant it could potentially select the
wrong width.

Fix this by not refining the width of physical or constant Vregs. Also
take into account the def Vreg when computing the width of the cmov.

Reviewed By: mxw

Differential Revision: D18127017

fbshipit-source-id: 430e00e9da0711bc4edebb7299829c76a823dc1c