Pass full FFP parser env to Rust
Summary:
A comment in rust_parser_ffi.ml states, "the less complicated structs we need to synchronize on the boundary between Rust and OCaml, the better." I agree!

Therefore, I want to suggest here that instead of defining the `parser_opts` type, manually converting it, and keeping that manual conversion in sync with the `ParserEnv` type, we automatically convert `Full_fidelity_parser_env.t` (stubbing out the `stats` field) and pull out the subset of fields we care about on the Rust side.

Reviewed By: losvald

Differential Revision: D18051143

fbshipit-source-id: 4b07811ffeb4cbd92b9d0ca190219c63b91e4998