Pull rust_parser_ffi parse function out of macro
Summary:
Having lots of logic in a macro context is inconvenient. We don't get formatting, linting, or proper IDE services within `macro_rules!`. Since this function is particularly long, it would be nice to get it out of the `parse!` macro.

To do this, I think it needs to be generic over the smart-constructors and smart-constructors-state types, so I exposed those types from each parser module.

Reviewed By: losvald

Differential Revision: D18052360

fbshipit-source-id: 61979907ae1a0f5a163c5056310f9574eddd4e93