Model HH-autoimport of funcs/consts as 'use {function, const}' statements
Summary:
As a result of past work to standardize HH-autoimporting between the typechecker and compiler, it is semantically equivalent to a series of `use {type, function, const, namespace}` statements at the top of every file and/or namespace block (except that they can be overridden by users in code).

This diff literally considers function, const, and namespace autoimports to be `use {function, const, namespace}` statements. When types are unified, it will allow us to kill all special-cased code for HH-autoimporting and instead re-use the existing elaboration machinery.

Differential Revision: D17847726

fbshipit-source-id: 76a740ddaad42f8287a4678e1d27dd49d6a917fd