Do not reserve rvmsp() in prologue and cross-trace ABIs
Summary:
There's nothing special about `rvmsp()` inside traces. It's used just for
passing inputs/outputs between some traces, which define that intent
explicitly. So, do not reserve it in prologues and cross-trace ABIs.

Without this diff, register allocator does not treat `rvmsp()` as a callee
saved register and it aggressively moves the content of `rvmsp()` to another
register. This causes unnecessary register pressure in the future prologues.

Reviewed By: ricklavoie

Differential Revision: D18132553

fbshipit-source-id: 9a33d57bfb889fe010371e80d1cb7275d87b183f