Stop generating inout wrappers
Summary:
Now that we're removing references from hhvm we no longer need to support by-reference wrappers for inout functions. This means that we no longer need to generate mangled function names for inout functions. It also means that we cannot rely on the reffy counterpart for inout functions in various places. This diff does the following:

 - Remove support for byRef() arguments on functions
 - Strip all inout-name-mangling related code from hhvm
 - Stop generating wrappers for inout methods and functions
 - Reuses the existing mechanism for enforcing reffiness to instead enforce that inout annotations are correct

Reviewed By: jano

Differential Revision: D17814910

fbshipit-source-id: 5dbcb9d972a9b18d80c1a8d4a040052e86b57687