Change the way parent is stored in genv
Summary: At the moment, parent ID (parent name) and parent type are stored as separate fields in `genv`, and the absence of parent is represented by setting them to `""` resp. `(Reason.none, Typing_defs.make_any ())`.  This feels wrong.  Why not represent parent as an optional pair of parent name and parent type, with `None` representing the absence of parent?  This is what this diff does.

Reviewed By: andrewjkennedy

Differential Revision: D18083497

fbshipit-source-id: c3ce9a31886f2839190950daf316bdb89a56835b