Set class parent when constructing class type checking environment
Summary: When constructing an environment for type checking a class, we set the parent ID based on the parent of the class but not the parent type.  The latter is set later in `Typing.class_def_`.  Shuffle things around a little, so that both parent ID and parent type are set when constructing the class type checking environment.

Reviewed By: andrewjkennedy

Differential Revision: D18083498

fbshipit-source-id: 80aced76fdbe6e8559d9bd6b696a1a533db6dece