Add a test showing the need of global_tpenv
Summary: My unsuccessful attempt to kill off `global_tpenv` (D18117725) revealed a test case that is missing from our test suite: type constant access inside the body of a lambda.  When a lambda is type checked, all the continuations (and associated `tpenv`s) are stashed, so if the constraints on generic type parameters that are currently added to `global_tpenv` are added to the local `tpenv` instead, they are lost when we are done type checking the lambda.

Reviewed By: andrewjkennedy

Differential Revision: D18169557

fbshipit-source-id: 64d53fba0e9593c156d04313f7f54eff8e280431