Fix another loop spill hoisting crash
Summary:
D18002123 fixed one bug, but caused another. If we remove a reload
within a loop because we've hoisted the spill, we might remove the
(only) definition of a Vreg. If the Vreg is used in a phi leaving the
loop, we'll leave a Vreg without any definition. This causes
assertions and various other crashes later on (depending on whether
we're in debug mode or not).

Rather than removing the reload, turn it into a copy instead. This has
the same effect, but keeps the Vreg definition. Later parts of the
register allocator will optimize away the copy as necessary. Do the
same for spill instructions for symmetry. Rematerialized instructions
can be turned into nops, which avoids the need to delete any
instructions at this point.

Reviewed By: mofarrell

Differential Revision: D18162467

fbshipit-source-id: 00b1031a130e129afab5c74ef683f3146b9ef2dc