Pass size index to create closure, and don't init props
Summary: This specializes createClosureRepoAuth in a similar way to how we do for ObjectData::newInstance.  The key here is that we pass in the size index from the JIT so that it does not need to be pulled off of the Class*.  This saves loading the class into cache.  As a bonus, we don't need to initialize the properties to uninit for a closure because the JIT is guaranteed to write them all immediately after (this is probably not that important given that most closures have very few properties).

Differential Revision: D18037967

fbshipit-source-id: 3a9f5d5b7323042a49f5b7c11a0a0abec0754298