Weaken exhaustiveness check for like enums
Summary:
The continuation code ignores enums when making exhaustive cases, stating that enum exhaustiveness is checked elsewhere. This makes the code also ignore like enums. This manifests in cases such as

```
enum MyEnum: int {
  A = 0;
  B = 1;
}

function f(~MyEnum $m): int {
  switch ($m) {
    case MyEnum::A:
      return 3;
    case MyEnum::B:
      return 4;
  }
}
// Error, the function implicitly returns void because the switch isn't exhaustive
```

Expanding the check to ~enums resolves the error. In theory, the error is correct, as the like enum could truly not be a value of the enum (HHVM does not enforce enums). In practice, this means that every switch on enums that come from function arguments or function calls will need default cases, which is wildly impractical and kills the utility of enum exhaustiveness checks.

Reviewed By: kmeht

Differential Revision: D18143620

fbshipit-source-id: 9623352e87fa47cd7f8a147dc3ffa7033d5343c7