Fix logic bugs related to const pointer types
Summary:
jit::Type uses a single union to store any constant values and any specialized types. It can represent "PtrToStaticArray(const value = ...)" and "PtrToStaticArray(kind = Mixed)", but not both.

However, the current logic for jit::Type::arrSpec / jit::Type::clsSpec always returns a specialization for constant types. For non-pointer types, that's valid, because the specialization is implied by the constant value itself. The same is not true for pointer types; in fact, constant pointer types may not even point to a valid value (an iterator "end" pointer is an example of such a type).

In this diff, we fix the logic around these types, and we explain how these cases can occur in a comment.

I tried to write a small test case for this bug, but even knowing how it works and turning on iterator specialization, I couldn't trigger a failure locally. However, I did run a small test case and confirm that the type that we settle on for an iterator position is PtrToStaticArray, without the specialization, based on load-elim's usage of the end to constrain the position.

Reviewed By: mxw

Differential Revision: D18121991

fbshipit-source-id: a236cf8b7ce8924262efc6332f1a46606ab3ffcd