Merge hint_to_ty and name_to_ty into node_to_ty.
Summary: We're building `ty`s from an increasing number of places, and want it to be as easy as possible.  Part of that is cutting down on unnecessary match statements, so if we combine `hint_to_ty` and `name_to_ty` we can move all of that switching into a single function.

Differential Revision: D18101999

fbshipit-source-id: 20f6b29e87e7aac36d574ac0b0a1885354fccda9