PAUSE Rpc: keep metadata about Paused state
Summary:
Keep the metadata from `env` at the time when `PAUSE true` is received.

The reason for this is so we can remember a little bit of metadata about where we were at when the user paused file change-driven global rechecks.

The key piece of metadata is the recheck ID.

Reviewed By: jewelpit

Differential Revision: D18164873

fbshipit-source-id: b2fa41bc22beda50e73312f2e0f1fd7e98b100a0