Cancel current recheck when user pauses global rechecks
Summary:
Global rechecks in response to file changes can be paused. If the user changes the state to `Paused` during an ongoing recheck, we should cancel that recheck, which is what this diff does.

The effect of `PAUSE true` during a recheck is that the recheck will be canceled, while the result of `PAUSE false` is that the client will wait for the recheck to be finished.

NOTE:
Interrupt handlers are currently set up and used during type checking. MultiWorker executor (MultiThreadedCall) selects worker file descriptors as well as some input channels from clients. If a client is sending an RPC over such a channel, the executor will call that channel's designated handler.

`ServerMain` sets up a few of these handlers, and the one we're interested in for this change is the __priority__ client interrupt handler. This handler is only listening to RPCs sent over the priority channel. The client decides which RPCs are sent over the priority channel vs. the default channel.

In the priority client interrupt handler, we actually handle the RPC that the client is sending and then returning the updated environment to the MultiWorker executor, along with the decision of whether it should continue executing or cancel. This means that we can examine the environment after handling the RPC to check whether the user paused file changes-driven global rechecks during the *current* recheck. If that is the case, then the current recheck should be canceled.

The reason we care about whether it's the current recheck that's paused or some other one is because global rechecks can still happen even if *file changes-driven* global rechecks are paused. This is because the user can explicitly request a recheck by calling an `hh_client` command that *requires* a full recheck. There are a number of such commands, but the most obvious one is just `hh` (defaults to `hh check`).

This change makes it possible to immediately set the server into paused mode AND cancel the current recheck. There are two cases I'm imagining the user doing this in:
1) The user notices that some change they made is taking a while to recheck, and so they want to cancel the recheck because they want to make further changes, or retry the recheck with the `--remote` argument
2) While the server is in the paused state, the user explicitly starts a full recheck, but then decides that they want to cancel it

In both cases, running `hh --pause` on the command line should stop the recheck if it's in the middle of the type checking phase.

Note that interrupts are only getting set up for the type checking phase, so if the server is in the middle of, say, the redecl phase, it's not going to be interrupted until it gets to the type checking itself. In some cases, redecling can be very costly. However, shallow redecl should be considerably less costly, therefore I'm not sure we need to support interrupting redecling.

Reviewed By: jewelpit

Differential Revision: D18172177

fbshipit-source-id: 4b1162729221586ca6cd5c4f6b3dacbc37a40e10