Strip HH\ prefix for autoimported names in Typing_reason
Summary:
This updates `Typing_reason` to strip the `HH\` prefix from autoimported names.

We do this because some extremely common names are autoimported, like Collections, Hack arrays, and Awaitables. As we unify the typechecker and compiler's understanding of autoimported types, we don't want the `HH\` prefix to appear in error messages and hover types, because it may confuse users into writing that in code.

Reviewed By: vassilmladenov

Differential Revision: D18076965

fbshipit-source-id: 7f5a72b6e195bde629485c50021e28d5760d5ea7