Don't generate declarations of built-in constants declared in .hhi files
Summary: We currently only avoid generating declarations for "pseudo-constants" (like `__FILE__` and `__CLASS__`) but there are other built-in constants, like `PHP_INT_MAX`, which are declared in `.hhi` files.  We must not generate declarations of those as doing so produces an error when type checking the output with `hh_single_type_check`.

Reviewed By: andrewjkennedy

Differential Revision: D18171538

fbshipit-source-id: a5086bc524bc1b767e6ec5be75d5d614cd258724