Fix optimization in PositionedToken::with_ methods
Summary: There's a comment here indicating that we are expecting these `make_mut` invocations to always avoid cloning the inner value, since we expect that the refcount is always 1. But I think we've actually guaranteed the opposite--by adding `Rc::clone`, it looks like we guarantee that the refcount is *more* than 1. The clone isn't necessary here--we can just omit it.

Reviewed By: shiqicao, losvald

Differential Revision: D18161689

fbshipit-source-id: 83c7bc3ce0f1aea0afe4bec12d7d8559be40f191