Make PositionedToken a type alias rather than a newtype
Summary: Making PositionedToken a newtype prevents us from taking advantage of RcOc's implementation of Deref (so we must write `self.0` or `token.0` everywhere) and requires us to expose Rc features when we want them (e.g. the `clone_rc` method). I'm not sure that we gain much from the wrapper type, so I'm suggesting that we remove it.

Reviewed By: dabek

Differential Revision: D18161686

fbshipit-source-id: f7dde01c267358801cbf0f23c2bd7eabbc4076d8